ADD:Double Cliffs
The only remaining non-flat cases are those where the bottom cell has cliffs on both sides.
This leaves all options open for the top edge.
It could be either flat, a slope, or a cliff.
We're only interested in the cliff-cliff-slope case, as it is the only one with terraces.