REF: Perturb Strength
It is now clear that we have distorted the grid, but the effect is very subtle.
There is at most a 1 unit adjustment in each dimension.
So the theoretical maximum displacement is √3 ≈ 1.73 units, which would be extremely rare, if it happens at all.
As our cells have an outer radius of ten units, the perturbations are relatively small.

The solution is to add a strength setting to HexMetrics so we can scale the perturbations. Let's try a strength of 5.