REF: Tweaking
We now have a fully correct perturbed mesh.
Its exact appearance depends on the specific noise, its scale, and the perturbation strengths.
In our case, the perturbation might be a bit too strong.
While an irregular appearance is nice, we don't want the cells to deviate too much from the regular grid.
After all, we still use that to determine which cell we're editing.
And if the size of cells varies too much, it will be harder to fit content inside them later.