REF: Removing Invalid Roads
We made sure that roads are only added when allowed.
Now we have to make sure to remove them when they become invalid later.

For example, when adding a river.
We could disallow rivers to be placed on top of roads, but rivers aren't stopped by roads.
Let them wash the roads away.