FIX: Rendering On Top
When drawing roads, you'll see road segments pop up in between cells.
The middle of these segments will be magenta, transitioning to blue at the sides.

However, when you move the camera around, the segments will probably flicker, and sometimes disappear completely.
This happens because the road triangles exactly overlap the terrain triangles.
It is arbitrary which ends up rendered on top.