ADD: Roads Across Cells
When triangulating rivers, we only had to deal with at most two river directions per cell.
We could identify the five possible scenarios and triangulate them differently to create well-behaved rivers.
However, there are fourteen possible scenarios for roads.
We're not going to use a different approach for each of these scenarios.
Instead, we'll treat each of the six cell directions the exact same way, regardless of the specific road configuration.

When there's a road running across a cell part, we'll run it straight to the cell center, without going outside of the triangular zone.
We'll draw a road segment from the edge halfway towards the center.
Then we'll use two triangles to cover the rest of the way to the center.