ADD: Roads with complete edges
We can now see the roads, but they taper toward the cell centers.
Because we're not checking which of the fourteen road scenarios we're dealing with, we can't move the road center to produce more pleasing shapes.
What we can do instead is add additional road edges in other parts of the cell.

When a cell has roads through it, but not in the current direction, add a road edge triangle.
This triangle is defined by the center, and the left and right middle vertices.
In this case, only the center vertex lies in the middle of the road. The other two vertices sits at its edge.