ADD: Smooth roads
Our roads are now complete. Unfortunately, this approach produces bulges at the center of cells.
Placing the left and right vertices halfway between the center and corners is fine, when there's a road adjacent to them.
But when there isn't, it results in a bulge.
To counter this, we could place the vertices closer to the center in those cases.
Specifically, by interpolating with ¼ instead of with ½.