ADD: Smarter Searching plus Priority Queue
Although our search algorithm finds the shortest path, it spends a lot of time investigating cells that obviously won't be part of that path.
Although A* is a good algorithm, our implementation of it is not that efficient. That's because we're using a list to store the frontier.
Our queue must support an enqueue and a dequeue operation, based on priority.